scripts/requirements.txt
pyannote.audio==3.1.1
torch>=2.0.0
librosa>=0.10.0
numpy>=1.21.0
soundfile>=0.12.0
scipy>=1.9.0
scikit-learn>=1.1.0
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.20.0
